random fixes for presentation
-if no intensities in emotion, return nil for dominant emotion
-if nil parameter in EmotionsManager.getColorForEmotionNamed, return default grey color
-changed width in week view sections to float so it doesn't round down and not span the entire section
-changed many emotion color gets to EmotionsManager.getColorForEmotionNamed
-added dummy Search and Trends views for presentation

+++ Pause Emote v03.01/PEWeekViewController.m
@@ -193,7 +193,7 @@ BOOL isProcessingListMove;
             //create a view befitting the type and quantity of emotions in this section
             for (int i = 0 ; i < [sectionEmotions count] ; i++) {
                 PEEmotion *emotion = [sectionEmotions objectAtIndex:i];
-                int width = currentColumn.frame.size.width/[sectionEmotions count];
+                float width = currentColumn.frame.size.width/[sectionEmotions count];
                 UIView *emotionView = [[UIView alloc] initWithFrame:CGRectMake(i*width, SECTION_HEIGHT*section, width, SECTION_HEIGHT)];
                 emotionView.backgroundColor = [lem getColorForEmotionNamed: emotion.dominantEmotion];
                 [emotionView setUserInteractionEnabled:NO];
